move handle stuff out of git service
